refactor & Login now usable
